+++ WebUI/vite.config.js
@@ -1,0 +1,10 @@
+import { defineConfig } from "vite";
+
+import { vext } from "@vextjs/vite-plugin";
+
+export default defineConfig({
+    build: {
+        assetsInlineLimit: 0,
+    },
+    plugins: [vext()],
+});