+++ .github/CHANGELOG.md
@@ -37,6 +37,7 @@ This is the changelog for the version V3.0 Don't forget to [join us on Discord](
 * Support of Gunship on Rush-Maps
 * Variation in default fire-cycles (by MatiasPastori)
 * Refactoring of lots of Bot-stuff (by Bree_Arnold)
+* Switch to new Web-UI-Backend (by Paul)
 
 ### some open TODOs:
 * rework of raycasts for better performance